improve image installer messages, idempotency and cleanup

+++ data/download-images.sh
@@ -16,7 +16,7 @@ if [ ! -f ./imagenet-val/n01440764/ILSVRC2012_val_00006697.JPEG ]; then
   mkdir -p imagenet-val
   mkdir -p ILSVRC2012_work/imagenet-val
   pushd ILSVRC2012_work/imagenet-val >/dev/null
-  echo extracting ILSVRC2012_img_val.tar
+  echo "Extracting ILSVRC2012_img_val.tar..."
   tar -xf ../../ILSVRC2012_img_val.tar
   echo "Moving images to imagenet-val/<synset_id>..."
   wget -q -O valprep.sh https://raw.githubusercontent.com/soumith/imagenetloader.torch/master/valprep.sh
@@ -32,10 +32,11 @@ fi
 if [ ! -f ./imagenet-val/n99991003/20260310_135804.jpg ]; then
   echo "Downloading custom validation images..."
   f=devboard-validation-images.tar
-  wget -O $f https://downloads.iotconnect.io/ai-data/mobilenet-v2/$f
+  if [ ! -f $f ]; then
+    wget -O $f https://downloads.iotconnect.io/ai-data/mobilenet-v2/$f
+  fi
   echo "Extracting validation images..."
   tar -xf $f -C ./imagenet-val
-  rm -f $f
 else
   echo "Found imagenet-val/n99991003/20260310_135804.jpg. Assuming custom validation images are present and skipping."
 fi
@@ -43,28 +44,32 @@ fi
 if [ ! -f train/n99991002/20260311_123929.jpg ]; then
   echo "Downloading custom training images..."
   f=devboard-training-images.tar
-  wget -O $f https://downloads.iotconnect.io/ai-data/mobilenet-v2/$f
+  if [ ! -f $f ]; then
+    wget -O $f https://downloads.iotconnect.io/ai-data/mobilenet-v2/$f
+  fi
   mkdir -p train
   echo "Extracting training images..."
   tar -xf $f -C ./train
-  rm -f $f
 else
   echo "Found train/n99991002/20260311_123929.jpg. Assuming custom training images are present and skipping."
 fi
 
 if [ ! -f calibration.npz ]; then
   echo "Installing requirements and generating calibration data..."
+  rm -rf venv-calib # in case it's left over
   python3 -m venv venv-calib
   source venv-calib/bin/activate
   python3 -m pip install --quiet -r requirements.txt
   python3 generate-representative-dataset.py
   deactivate
-  rm -rf venv-calib
 else
   echo "Found calibration.npz, skipping calibration data generation."
 fi
 
 # only after everything was successful:
+echo "Cleaning up..."
+rm -f devboard-validation-images.tar
+rm -f devboard-training-images.tar
 rm -f ILSVRC2012_img_val.tar
 
 echo "Done."
